+++ WorQit/WorQit/FilterPage.xaml
@@ -32,14 +32,13 @@
         </Grid.RowDefinitions>
         <Image x:Name="image" Grid.Column="2" Source="Assets/logo WorQit.png" Grid.ColumnSpan="2"/>
         <TextBlock Grid.Column="3" Grid.Row="1" TextWrapping="Wrap" Text="Filters aanpassen" FontSize="16" Foreground="#FF195695" HorizontalAlignment="Center" VerticalAlignment="Center" Margin="17,6,17,5" Height="21" Width="120"/>
-        <TextBlock Grid.Column="1" Grid.Row="3" TextWrapping="Wrap" Text="Afstand" Grid.ColumnSpan="2" VerticalAlignment="Center" Height="20" Margin="0,6"/>
+        <TextBlock Grid.Column="1" Grid.Row="3" TextWrapping="Wrap" Text="max. afstand" Grid.ColumnSpan="2" VerticalAlignment="Center" Height="20" Margin="0,6"/>
         <TextBox x:Name="txtDistance" Grid.Column="3" Grid.Row="3" TextWrapping="Wrap" Text="" Grid.ColumnSpan="2"/>
-        <TextBlock Grid.Column="1" Grid.Row="4" TextWrapping="Wrap" Text="Salaris" Grid.ColumnSpan="2" VerticalAlignment="Center" Margin="0,6" Height="20"/>
+        <TextBlock Grid.Column="1" Grid.Row="4" TextWrapping="Wrap" Text="min. salaris" Grid.ColumnSpan="2" VerticalAlignment="Center" Margin="0,6" Height="20"/>
         <TextBox x:Name="txtSalary" Grid.Column="3" Grid.Row="4" TextWrapping="Wrap" Text="" Grid.ColumnSpan="2"/>
-        <TextBlock Grid.Column="1" Grid.Row="5" TextWrapping="Wrap" Text="Uren" Grid.ColumnSpan="2" VerticalAlignment="Center" Margin="0,6" Height="20"/>
+        <TextBlock Grid.Column="1" Grid.Row="5" TextWrapping="Wrap" Text="min. uren" Grid.ColumnSpan="2" VerticalAlignment="Center" Margin="0,6" Height="20"/>
         <TextBox x:Name="txtHours" Grid.Column="3" Grid.Row="5" TextWrapping="Wrap" Text="" Grid.ColumnSpan="2"/>
-        <TextBlock Grid.Column="1" Grid.Row="6" TextWrapping="Wrap" Text="Functie" Grid.ColumnSpan="2" VerticalAlignment="Center" Margin="0,12,0,0" Height="20"/>
-        <TextBox x:Name="txtFunction" Grid.Column="3" Grid.Row="6" TextWrapping="Wrap" Text="" Grid.ColumnSpan="2"/>
-        <Button x:Name="btnBack" Content="terug" Grid.Column="3" Grid.Row="11" Style="{StaticResource ButtonStyle}" HorizontalAlignment="Right" Margin="0,89,0,88" Width="154" Click="btnBack_Click"/>
+        <Button x:Name="btnBack" Content="terug" Grid.Column="2" Grid.Row="8" Style="{StaticResource ButtonStyle}" Click="btnBack_Click" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Grid.ColumnSpan="2" Margin="0,2,0,0"/>
+        <Button x:Name="btnApply" Content="filters toevoegen" Grid.Column="2" Grid.Row="7" Style="{StaticResource ButtonStyle}" Click="btnApply_Click" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Grid.ColumnSpan="2" Margin="0,0,0,3"/>
     </Grid>
 </Page>
